luv-live-image: Append kernel parameter for handling crashes
Add the crashkernel parameter to indicate the kernel to reserve memory
for the kernel and initrd images that will be used in the event of a
crash.

TODO: Define different size of memory for different amounts of available
memory in the system. Also, explain why we are using 256M and not other

Signed-off-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>