kexec-tools: Add a recipe to utilize version 2.0.7
Due to the following change in the Linux kernel,

commit 62e32ac3505a0cab1c5ef8ea2c0eab3b26ed855f
Author: Joe Perches <joe@perches.com>
Date:   Wed Jul 31 13:53:47 2013 -0700

    printk: rename struct log to struct printk_log

Only version 2.0.7 or later of vmcore-dmesg may be used correctly
parse the dmesg buffer. Thus, in order to have tighter control
on the version, use a recipe that explicitly selects the desired version.

Also, this recipe is more inline with the approach used in the meta
layer and allows reuse of the common recipe kexec-tools.inc

Patches are removed because all of them are included in version 2.0.7

Signed-off-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>