luv-crash-handler: Save dmesg buffer to non-volatile media
The dmesg buffer can provide valuable information to determine
the causes of a kernel crash. When the systems boots to
runlevel 3, the luv-crash-handler uses vmcore-dmesg to retrieve
the dmesg buffer from the vmcore dump and saves it to the test
results partition of the LuvOS boot media.

This information can then be sent to the developers to investigate
further on the causes of the crash.

Signed-off-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>