initscripts: check luv.config to see if test suite needs to be ran
The file luv.config contains selections about which test suites were selected
for execution.

luv-test-manager will look for the name of the runner. If present it will
check if the word "yes" is present. If it is present, the test suite will
be executed.

Additionally, the presence of the luv.config file is expected. Thus, if the
file is missing, it is considered an error and we abort execution.

Signed-off-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>