fwts: get_nextvariable: copy only the needed name bytes
We trust the name_size passed by the user_space. Thus, there is not
need to copy from user more bytes than specified.

Furthermore, this behavior could mask potential problems in the
firmware: to start traversing the list of variable names, some firmware
implementations may just blindly check if name[0] is '\0' without
even checking if the size of the buffer passed by the operating system
is at least 2 bytes (UEFI variable names are stored in strings of 2-byte
characters). This behaviour is wrong as the firmware must not access
memory beyond the name buffer. By providing a 1024-bytes buffer of
zeroes we allow the name[0] == '\0' to pass, regardless of the size
of the buffer.

Signed-off-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>