runqemu: Add option for BIOS binary filename
It may be possible that the default name for the BIOS binary that
qemu uses change. For instance, it was called bios.bin for 1.5.0
and is bios-256k.bin for 2.1.0. If we intend to use a custom BIOS
binary file, it would be good to be able to specify the name of the
binary via the -bios option of qemu.

Signed-off-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>