shim-signed: add recipe for shim-signed
This recipe adds functionality to retrieve and install Matthew
Garret's shim signed by Microsoft. By using this signed binary,
there is not need to add our own signatures to the UEFI firmware.

However, we do need to generate a key and certificate to sign
binaries so that they are recognized by shim.

Signed-off-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>